LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/keyword-sizes-for-intrinsic-contributions-expected.txt


PASS .test 1
PASS .test 2
PASS .test 3
PASS .test 4
PASS .test 5
PASS .test 6
PASS .test 7
PASS .test 8
PASS .test 9
PASS .test 10
PASS .test 11
PASS .test 12
FAIL .test 13 assert_equals:
<div class="test min-width" data-expected-width="70"><div style="min-width: min-content">XXX XXX</div></div>
width expected 70 but got 10
FAIL .test 14 assert_equals:
<div class="test min-width" data-expected-width="70"><div style="min-width: fit-content">XXX XXX</div></div>
width expected 70 but got 10
FAIL .test 15 assert_equals:
<div class="test min-width" data-expected-width="150"><div style="min-width: max-content">XXX XXX</div></div>
width expected 150 but got 10
PASS .test 16
FAIL .test 17 assert_equals:
<div class="test min-width" data-expected-width="70"><div style="min-width: min-content">XXX XXX</div></div>
width expected 70 but got 10
FAIL .test 18 assert_equals:
<div class="test min-width" data-expected-width="90"><div style="min-width: fit-content">XXX XXX</div></div>
width expected 90 but got 10
FAIL .test 19 assert_equals:
<div class="test min-width" data-expected-width="150"><div style="min-width: max-content">XXX XXX</div></div>
width expected 150 but got 10
PASS .test 20
FAIL .test 21 assert_equals:
<div class="test min-width" data-expected-width="70"><div style="min-width: min-content">XXX XXX</div></div>
width expected 70 but got 10
FAIL .test 22 assert_equals:
<div class="test min-width" data-expected-width="150"><div style="min-width: fit-content">XXX XXX</div></div>
width expected 150 but got 10
FAIL .test 23 assert_equals:
<div class="test min-width" data-expected-width="150"><div style="min-width: max-content">XXX XXX</div></div>
width expected 150 but got 10
PASS .test 24
PASS .test 25
FAIL .test 26 assert_equals:
<div class="test max-width" data-expected-width="70"><div style="max-width: fit-content">XXX XXX</div></div>
width expected 70 but got 210
FAIL .test 27 assert_equals:
<div class="test max-width" data-expected-width="150"><div style="max-width: max-content">XXX XXX</div></div>
width expected 150 but got 210
PASS .test 28
PASS .test 29
FAIL .test 30 assert_equals:
<div class="test max-width" data-expected-width="90"><div style="max-width: fit-content">XXX XXX</div></div>
width expected 90 but got 210
FAIL .test 31 assert_equals:
<div class="test max-width" data-expected-width="150"><div style="max-width: max-content">XXX XXX</div></div>
width expected 150 but got 210
PASS .test 32
PASS .test 33
FAIL .test 34 assert_equals:
<div class="test max-width" data-expected-width="150"><div style="max-width: fit-content">XXX XXX</div></div>
width expected 150 but got 210
FAIL .test 35 assert_equals:
<div class="test max-width" data-expected-width="150"><div style="max-width: max-content">XXX XXX</div></div>
width expected 150 but got 210
PASS .test 36
XXX XXX
XXX XXX
XXX XXX
XXX XXX

XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX

XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX

XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
